Refactored Elasticsearch Functionality
This commit refactors some of the Elasticsearch querying functionality and sets up all of the new dependencies for injection.
