+++ api/Doyen.API/Doyen.API/Experts/Elasticsearch/Http/IElasticsearchHttpClient.cs
@@ -1,0 +1,7 @@
+﻿namespace Doyen.API.Experts.Elasticsearch.Http
+{
+    public interface IElasticsearchHttpClient
+    {
+        Task<string> SendSearchPostRequest(string requestBody);
+    }
+}
