+++ api/Doyen.API/Doyen.API/Experts/Elasticsearch/PubMedPublications.cs
@@ -1,4 +1,4 @@
-﻿namespace QuickType
+﻿namespace Doyen.API.Experts.Elasticsearch
 {
     using Newtonsoft.Json;
     using Newtonsoft.Json.Converters;
@@ -249,7 +249,7 @@
             if (reader.TokenType == JsonToken.Null) return null;
             var value = serializer.Deserialize<string>(reader);
             long l;
-            if (Int64.TryParse(value, out l))
+            if (long.TryParse(value, out l))
             {
                 return l;
             }
