+++ api/Doyen.API/Doyen.API/Experts/Elasticsearch/Settings/ElasticsearchSettings.cs
@@ -1,4 +1,4 @@
-﻿namespace Doyen.API.Elasticsearch.Settings
+﻿namespace Doyen.API.Experts.Elasticsearch.Settings
 {
     public class ElasticsearchSettings : IElasticsearchSettings
     {
